+++ deep8/nodes/Linear.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_LINEAR_H
 #define DEEP8_LINEAR_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
 /**
@@ -41,32 +43,16 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the Linear Function needs only 1 input");
-
-        this->outputShape = this->inputs[0]->outputShape;
-    }
+	void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-        auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
 
-        eTVec(output).device(*device) = eTVec(inputs[0]) * a + b;
-    }
-
-    void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-                     const Tensor<T> *output,
-                     const Tensor<T> *outputGradient,
-                     size_t index,
-                     Tensor<T> *iGradient) override {
-        DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
-
-        auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
-
-        eTVec(iGradient).device(*device) += eTVec(outputGradient) * a;
-    }
+	void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
+					const Tensor<T> *output,
+					const Tensor<T> *outputGradient,
+					size_t index,
+					Tensor<T> *iGradient) override;
 
 
 #ifdef HAVE_CUDA
