+++ deep8/nodes/Log.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_LOG_H
 #define DEEP8_LOG_H
 
+#include "Function.h"
+
 
 namespace Deep8 {
 
@@ -47,32 +49,16 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the Log Function needs only 1 input");
-
-        this->outputShape = this->inputs[0]->outputShape;
-    }
+	void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-        auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
 
-        eTVec(output).device(*device) = eTVec(inputs[0]).log();
-    }
-
-    void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-                     const Tensor<T> *output,
-                     const Tensor<T> *outputGradient,
-                     size_t index,
-                     Tensor<T> *iGradient) override {
-        DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
-
-        auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
-
-        eTVec(iGradient).device(*device) += eTVec(outputGradient).binaryExpr(eTVec(inputs[0]), LogBackwardExpr<T>());
-    }
+	void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
+					const Tensor<T> *output,
+					const Tensor<T> *outputGradient,
+					size_t index,
+					Tensor<T> *iGradient) override;
 
 
 #ifdef HAVE_CUDA
