+++ deep8/nodes/MinusScalar.cpp
@@ -1,0 +1,43 @@
+#include "MinusScalar.h"
+
+namespace Deep8 {
+
+template <typename T>
+void MinusScalar<T>::check() {
+	Function<T>::check();
+
+	DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the inputs size must be 1 in MinusScalar Function");
+
+	this->outputShape = this->inputs[0]->outputShape;
+}
+
+template <typename T>
+void MinusScalar<T>::forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) {
+	auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
+
+	eTVec(output).device(*device) = eTVec(inputs[0]) - scalar;
+}
+
+#ifdef HAVE_HALF
+template <>
+void MinusScalar<half>::forwardCPU<half>(const std::vector<const Tensor<half>*> &inputs, Tensor<half> *output) {
+	DEEP8_RUNTIME_ERROR("CPU not support half");
+}
+#endif // HAVE_HALF
+
+template <typename T>
+void MinusScalar<T>::backwardCPU(const std::vector<const Tensor<T>*> &inputs,
+								const Tensor<T> *output,
+								const Tensor<T> *outputGradient,
+								size_t index,
+								Tensor<T> *iGradient) {
+	auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
+
+	DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
+
+	eTVec(iGradient).device(*device) += eTVec(outputGradient);
+}
+
+DEEP8_DECLARATION_INSTANCE(MinusScalar)
+
+}
