+++ deep8/nodes/MinusScalar.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_MINUSSCALAR_H
 #define DEEP8_MINUSSCALAR_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
 /*****************************************************************************************************************/
@@ -40,46 +42,16 @@ public:
 		check();
 	}
 
-	void check() override {
-		Function<T>::check();
-
-		DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the inputs size must be 1 in MinusScalar Function");
-
-		this->outputShape = this->inputs[0]->outputShape;
-	}
+	void check() override;
 
 protected:
-	template <typename real>
-	void forwardCPUImpl(const std::vector<const Tensor<real>*> &inputs, Tensor<real> *output) {
-		auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-		eTVec(output).device(*device) = eTVec(inputs[0]) - scalar;
-	}
-
-#ifdef HAVE_HALF
-	template <>
-	void forwardCPUImpl<half>(const std::vector<const Tensor<half>*> &inputs, Tensor<half> *output) {
-		DEEP8_RUNTIME_ERROR("CPU not support half");
-	}
-#endif // HAVE_HALF
-
-
-	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-		forwardCPUImpl(inputs, output);
-	}
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
 
 	void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-		const Tensor<T> *output,
-		const Tensor<T> *outputGradient,
-		size_t index,
-		Tensor<T> *iGradient) override {
-		auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
-
-		DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
-
-		eTVec(iGradient).device(*device) += eTVec(outputGradient);
-	}
-
+					const Tensor<T> *output,
+					const Tensor<T> *outputGradient,
+					size_t index,
+					Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
 	template <typename real>
