+++ deep8/nodes/MultiplyScalar.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_MULTIPLYSCALAR_H
 #define DEEP8_MULTIPLYSCALAR_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
 /******************************************************************************************/
@@ -40,32 +42,11 @@ public:
 		check();
 	}
 
-	void check() override {
-		Function<T>::check();
-
-		DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the inputs size must be 1 in MultiplyScalar Function");
-
-		this->outputShape = this->inputs[0]->outputShape;
-	}
+	void check() override;
 
 protected:
-	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-		auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-		eTVec(output).device(*device) = eTVec(inputs[0]) * scalar;
-	}
-
-	void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-					 const Tensor<T> *output,
-					 const Tensor<T> *outputGradient,
-					 size_t index,
-					 Tensor<T> *iGradient) override {
-		auto device = static_cast<CPUDevice*>(outputGradient->device())->eigenDevice;
-
-		DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
-
-		eTVec(iGradient).device(*device) += eTVec(outputGradient) * scalar;
-	}
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
+	void backwardCPU(const std::vector<const Tensor<T>*> &inputs, const Tensor<T> *output, const Tensor<T> *outputGradient, size_t index, Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
 	template <typename real>
