+++ deep8/nodes/ReShape.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_RESHAPE_H
 #define DEEP8_RESHAPE_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
 template <typename T>
@@ -23,65 +25,29 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the input size must be 1");
-        DEEP8_ARGUMENT_CHECK(reShape.nDims() < MAX_TENSOR_DIMS, "the reShape is error");
-        DEEP8_ARGUMENT_CHECK(this->inputs[0]->outputShape.size() == reShape.size(), "the input's Shape size must be equal to reShape size");
-
-        this->outputShape = reShape;
-    }
+	void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
+    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {}
 
-    }
-
-    void forwardGPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-
-    }
+    void forwardGPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {}
 
     void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
                      const Tensor<T> *output,
                      const Tensor<T> *outputGradient,
                      size_t index,
-                     Tensor<T> *iGradient) override {
-
-    }
+                     Tensor<T> *iGradient) override {}
 
     void backwardGPU(const std::vector<const Tensor<T>*> &inputs,
                      const Tensor<T> *output,
                      const Tensor<T> *outputGradient,
                      size_t index,
-                     Tensor<T> *iGradient) override {
-    }
+                     Tensor<T> *iGradient) override {}
 
 public:
-    void forward() override {
-		DEEP8_ARGUMENT_CHECK(NodeType::Variable == this->output->type, "the output must be Variable type");
-
-		auto x = static_cast<Variable<T>*>(this->inputs[0]);
-		auto y = static_cast<Variable<T>*>(this->output);
-
-		y->outputShape = this->outputShape;
-		
-		y->updateGradient = x->updateGradient;
-		y->value.storage  = x->value.storage;
-		y->value.offset   = x->value.offset;
-		y->value.shape    = this->outputShape;
-
-		if (x->updateGradient) {
-			y->gradient.storage = x->gradient.storage;
-			y->gradient.offset  = x->gradient.offset;
-			y->gradient.shape   = this->outputShape;
-		}
-    }
-
-    void backward() override {
-    }
+	void forward() override;
+	void backward() override;
 };
-
 
 }
