+++ deep8/nodes/ScalarMinus.h
@@ -1,6 +1,8 @@
 #ifndef DEEP8_SCALARMINUS_H
 #define DEEP8_SCALARMINUS_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
 /*****************************************************************************************************************/
@@ -41,33 +43,17 @@ public:
 		check();
 	}
 
-	void check() override {
-		Function<T>::check();
-
-		DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the inputs size must be 1 in ScalarMinus Function");
-
-		this->outputShape = this->inputs[0]->outputShape;
-	}
+	void check() override;
 
 protected:
-	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-		auto device = static_cast<CPUDevice*>(output->device)->eigenDevice;
-
-		eTVec(output).device(*device) = -eTVec(inputs[0]) + scalar;
-	}
-
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
 
 	void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
 		const Tensor<T> *output,
 		const Tensor<T> *outputGradient,
 		size_t index,
-		Tensor<T> *iGradient) override {
-		auto device = static_cast<CPUDevice*>(outputGradient->device)->eigenDevice;
+		Tensor<T> *iGradient) override;
 
-		DEEP8_ARGUMENT_CHECK(0 == index, "the index is error");
-
-		eTVec(iGradient).device(*device) -= eTVec(outputGradient);
-	}
 #ifdef HAVE_CUDA
 
 	template<typename real>
