+++ deep8/nodes/TanH.h
@@ -1,21 +1,11 @@
 #ifndef DEEP8_TANH_H
 #define DEEP8_TANH_H
 
+#include "Function.h"
+
 namespace Deep8 {
 
-template <typename T>
-struct TanHForwardExpr {
-    inline T operator()(T in) const {
-        return tanh(in);
-    }
-};
 
-template <typename T>
-struct TanHBackwardExpr {
-    inline T operator()(T outputGrad, T output) const {
-        return outputGrad * (T(1.0) - output * output);
-    }
-};
 
 
 #ifdef HAVE_CUDA
@@ -49,36 +39,11 @@ public:
         check();
     }
 
-    void check() override {
-        Function<T>::check();
-
-        DEEP8_ARGUMENT_CHECK(1 == this->inputs.size(), "the TanH Function needs only 1 input");
-
-        this->outputShape = this->inputs[0]->outputShape;
-    }
+	void check() override;
 
 protected:
-    void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override {
-		auto device = static_cast<CPUDevice*>(output->device())->eigenDevice;
-
-        eTVec(output).device(*device) = eTVec(inputs[0]).unaryExpr(TanHForwardExpr<T>());
-    }
-
-    
-    void backwardCPU(const std::vector<const Tensor<T>*> &inputs,
-					 const Tensor<T> *output,
-					 const Tensor<T> *outputGradient,
-					 size_t index,
-					 Tensor<T> *iGradient) override {
-        if (0 != index) {
-            DEEP8_RUNTIME_ERROR("the index of TanH backwardCPU is error");
-        }
-
-		auto device = static_cast<CPUDevice*>(iGradient->device())->eigenDevice;
-
-        eTVec(iGradient).device(*device) += eTVec(outputGradient).binaryExpr(eTVec(output), TanHBackwardExpr<T>());
-    }
-
+	void forwardCPU(const std::vector<const Tensor<T>*> &inputs, Tensor<T> *output) override;
+	void backwardCPU(const std::vector<const Tensor<T>*> &inputs, const Tensor<T> *output, const Tensor<T> *outputGradient, size_t index, Tensor<T> *iGradient) override;
 
 #ifdef HAVE_CUDA
